Plot histogram of structure counts
I added code to plot a histogram of the structure counts (aka, plot a histogram of the vector called "abinter.") We would like to make the bin size a user-defined input on the command line, with the default as 100. We would also need "abinter" (list of nucleotides which pass both cutoffs) as a text output (structure_counts.csv) and the histogram as a figure output (structure_counts_histogram.tiff).